+++ examples/rp/src/bin/usb_midi.rs
@@ -15,6 +15,7 @@ use embassy_rp::usb::{Driver, Instance, InterruptHandler};
 use embassy_usb::class::midi::{MidiClass, MidiClassConfig};
 use embassy_usb::driver::EndpointError;
 use embassy_usb::{Builder, Config};
+use embassy_rp::uid::uid_hex;
 use panic_probe as _;
 
 bind_interrupts!(struct Irqs {
@@ -34,7 +35,7 @@ async fn main(_spawner: Spawner) {
     let mut config = Config::new(0xc0de, 0xcafe);
     config.manufacturer = Some("Embassy");
     config.product = Some("USB-MIDI example");
-    config.serial_number = Some("12345678");
+    config.serial_number = Some(uid_hex());
     config.max_power = 100;
     config.max_packet_size_0 = 64;